+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Removed `get()` and added it to README to remove dependency from `lodash.get`
 
 ## 1.0.3
 - Simplified README to understand the features